pyproject.toml
[project]
name = "audio-to-daw"
version = "0.1.0"
description = "Tool to automatically reconstruct a DAW project from a single audio track."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.10.1",
    "structlog>=25.4.0",
    "python-dotenv>=1.1.1",
    "pedalboard>=0.9.18",
    "ipykernel>=6.30.1",
    "sounddevice>=0.5.2",
    "tqdm>=4.67.1",
    "matplotlib>=3.10.6",
    "pydub>=0.25.1",
    "loguru>=0.7.3",
    "h5py>=3.14.0",
    "hdf5plugin>=5.1.0",
    "librosa>=0.11.0",
    "pyloudnorm>=0.1.1",
    "mido>=1.3.3",
    "pyyaml>=6.0",
    "click>=8.0",
    "soundfile>=0.12.0",
    "scipy>=1.9.0",
    "torch>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.3",
    "isort>=6.0.1",
    "pyright>=1.1.403",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "F", "E"]

[tool.ruff.lint.isort]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"

[tool.black]
line-length = 120
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 120
